% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greets.R
\name{greets}
\alias{greets}
\title{A nice hello from Bukrarest}
\usage{
greets(name = "everybody")
}
\arguments{
\item{name}{specify who we want to greet.}
}
\value{
\code{NULL}
}
\description{
\code{\link[=greets]{greets()}} gives a \strong{warm} greeting from the Uros19.
}
\examples{
greets()
}
