% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urosdata.R
\name{urosdata}
\alias{urosdata}
\title{urosdata}
\usage{
data(urosdata)
}
\description{
Some testdata for the urosconf pkg with this two variables:
\itemize{
\item \code{v}: a factor of lower case letter
\item \code{x} random values from a standard normal distribution
}
}
